Cleaning out unused files
