+++ src/components/App.js
@@ -4,30 +4,52 @@ import Contacts from './ContactList'
 import NewContact from './NewContact'
 
 
-// Creating a simple div header
+const generateId = () => Math.round(Math.random() * 100000000);
 
-class App extends React.Component {
-  constructor (props) {
-    super(props);
 
-    this.handleClick = this.handleClick.bind(this);
+class App extends Component {
+  constructor () {
+    super();
+
+
+    this.state: = {
+
+      contacts: [
+          {
+            id: generateId(),
+            name: "Albert Einstein",
+            image_url: "https://i.pinimg.com/236x/95/8f/7a/958f7acae61910da05f327a38eb63530--innovation-design-imagination.jpg",
+            email: "aeinstein@example.com",
+            phone: "15555555555"
+          },
+          {
+            id: generateId(),
+            name: "Homer Simpson",
+            image_url: "https://upload.wikimedia.org/wikipedia/en/0/02/Homer_Simpson_2006.png",
+            email: "homer@compuglobalhypermeganet.com",
+            phone: "9395550113"
+          },
+          {
+            id: generateId(),
+            name: "Ghostbusters Inc.",
+            image_url: "https://upload.wikimedia.org/wikipedia/en/e/ee/Ghostbusters_logo.svg",
+            email: "pvenkmann@ghostbusters.com",
+            phone: "1234567890"
+          }
+        ]
+
+
+    }
+
   }
 
-  handleClick () {
-    // This needs to link to NewContact but I don't know how
-    window.alert('CLICK');
-  }
-
-  addContact (contact) {
-    console.log(contact)
-  }
 
   render() {
     return (
       <div className="row">
         <div className="Contacts-main col-md-6 offset-md-3">
           <header className="Contacts-header">
-            <h1 className="Contacts-title">Contacts  <button onClick={this.handleClick}>Add Contact</button></h1>
+            <h1 className="Contacts-title">Contacts</h1> <button>Add Contact</button> //TODO Link to NewContact
           </header>
             <Contacts />
         </div>
